Fix a "run as user" bug
Before, the UI would launch as the custom user "nobody_99_100", but then it would get the homedir of user 99, which
aliased to "nobody". So the user prefs were being stored in /nobody/.java/.userPrefs/... This version of runas.sh
re-uses the existing user if its UID and GID is appropriate. In this case, if the user specifies 99 and 100, the
container will properly run FileBot CLI and UI as "nobody".